source/requirements.txt
PyCryptodome
Flask
flask-wtf
flask-bootstrap
flask-migrate
flask-sqlalchemy
Flask-Login